pyproject.toml
[tool.poetry]
package-mode = false
name = "ruyah"
version = "0.1.0"
description = "Ru'yah serves as a comprehensive task management solution, designed to empower individuals and teams in efficiently navigating their daily responsibilities."
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ruyah" }]

[tool.poetry.dependencies]
python = "3.10"
django = "^5.1.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-redesign = "^0.4.0"
huggingface-hub = "^0.25.1"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
